pyproject.toml
[project]
name = "pysmile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pyrefly>=0.29.2",
    "ruff>=0.12.10",
]
